+++ apps/manager/src/shared/components/SectionModule/components/AddSection/SectionIDField.tsx
@@ -3,12 +3,12 @@ import i18n from "@dhis2/d2-i18n";
 import { RHFTextInputField } from "@hisptz/dhis2-ui";
 import React, { useEffect } from "react";
 import { kebabCase } from "lodash";
-import { VisualizationGroup } from "@packages/shared/schemas";
+import { Section } from "@packages/shared/schemas";
 
 export function SectionIDField() {
-	const { setValue } = useFormContext<VisualizationGroup>();
-	const label = useWatch<VisualizationGroup, "title">({
-		name: "title",
+	const { setValue } = useFormContext<Section>();
+	const label = useWatch<Section, "label">({
+		name: "label",
 	});
 
 	useEffect(() => {