+++ packages/shared/src/schemas/visualization.ts
@@ -1,4 +1,5 @@
 import { z } from "zod";
+import { LegendSet } from "@hisptz/dhis2-utils";
 
 export const visualizationFields = [
 	"access",
@@ -7,12 +8,12 @@ export const visualizationFields = [
 	"colSubTotals",
 	"colTotals",
 	"colorSet",
-	"columns[dimension,filter,legendSet[id,name,displayName,displayShortName,legends[id,color,startValue,endValue]],items[dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access]]",
+	"columns[dimension,filter,items[dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access]]",
 	"completedOnly",
 	"created",
 	"cumulative",
 	"cumulativeValues",
-	"dataDimensionItems[dataDimensionItemType,expressionDimensionItem[dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access],dataElement[dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access],dataElementOperand[dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access],reportingRate[dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access],programAttribute[dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access],programIndicator[dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access],indicator[dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access],programDataElement[dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access]]",
+	"dataDimensionItems[dataDimensionItemType,expressionDimensionItem[legendSet[id,name,displayName,displayShortName,legends[id,color,startValue,endValue]],dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access],dataElement[legendSet[id,name,displayName,displayShortName,legends[id,color,startValue,endValue]],dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access],dataElementOperand[legendSet[id,name,displayName,displayShortName,legends[id,color,startValue,endValue]],dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access],reportingRate[legendSet[id,name,displayName,displayShortName,legends[id,color,startValue,endValue]],dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access],programAttribute[legendSet[id,name,displayName,displayShortName,legends[id,color,startValue,endValue]],dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access],programIndicator[legendSet[id,name,displayName,displayShortName,legends[id,color,startValue,endValue]],dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access],indicator[legendSet[id,name,displayName,displayShortName,legends[id,color,startValue,endValue]],dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access],programDataElement[legendSet[id,name,displayName,displayShortName,legends[id,color,startValue,endValue]],dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access]]",
 	"description",
 	"digitGroupSeparator",
 	"displayDensity",
@@ -23,7 +24,7 @@ export const visualizationFields = [
 	"displayTitle",
 	"favorite",
 	"favorites",
-	"filters[dimension,filter,legendSet[id,name,displayName,displayShortName,legends[id,color,startValue,endValue]],items[dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access]]",
+	"filters[dimension,filter,items[dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access]]",
 	"fixColumnHeaders",
 	"fixRowHeaders",
 	"fontSize",
@@ -39,7 +40,7 @@ export const visualizationFields = [
 	"interpretations[id,created]",
 	"lastUpdated",
 	"lastUpdatedBy",
-	"legend[showKey,style,strategy,set[id,name,displayName,displayShortName]]",
+	"legend[showKey,style,strategy,set[id,name,displayName,displayShortName,legends[id,color,startValue,endValue,name]]]",
 	"measureCriteria",
 	"name",
 	"noSpaceBetweenColumns",
@@ -53,7 +54,7 @@ export const visualizationFields = [
 	"reportingParams",
 	"rowSubTotals",
 	"rowTotals",
-	"rows[dimension,filter,legendSet[id,name,displayName,displayShortName,legends[id,color,startValue,endValue]],items[dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access]]",
+	"rows[dimension,filter,items[dimensionItem~rename(id),name,displayName,displayShortName,dimensionItemType,expression,access]]",
 	"series",
 	"seriesKey",
 	"shortName",
@@ -75,6 +76,7 @@ export const visualizationFields = [
 	"userAccesses",
 	"userGroupAccesses",
 	"yearlySeries",
+	"relativePeriods",
 	"!attributeDimensions",
 	"!attributeValues",
 	"!category",
@@ -92,7 +94,6 @@ export const visualizationFields = [
 	"!organisationUnits",
 	"!periods",
 	"!programIndicatorDimensions",
-	"!relativePeriods",
 	"!rowDimensions",
 	"!userOrganisationUnit",
 	"!userOrganisationUnitChildren",
@@ -141,6 +142,10 @@ export enum DataDimensionItemType {
 	PROGRAM_INDICATOR = "PROGRAM_INDICATOR",
 }
 
+export type LegendSetConfig =
+	| LegendSet
+	| { dataItem: string; legendSet: LegendSet }[];
+
 export enum DigitGroupSeparator {
 	SPACE = "SPACE",
 }
@@ -175,7 +180,7 @@ const legendSet = z.object({
 	id: z.string(),
 	name: z.string(),
 	displayName: z.string(),
-	displayShortName: z.string(),
+	displayShortName: z.string().optional(),
 	legends: z.array(
 		z.object({
 			id: z.string(),
@@ -205,6 +210,7 @@ const dataDimensionItem = z.object({
 	expression: z.string().optional(),
 	access: accessSchema,
 	dimensionItemType: z.string(),
+	legendSet: legendSet.optional(),
 });
 
 export type DataDimensionItem = z.infer<typeof dataDimensionItem>;
@@ -217,6 +223,11 @@ const dimension = z.object({
 });
 
 export type DimensionConfig = z.infer<typeof dimension>;
+
+export enum LegendStrategy {
+	FIXED = "FIXED",
+	BY_DATA_ITEM = "BY_DATA_ITEM",
+}
 
 export const visualizationSchema = z.object({
 	access: accessSchema,
@@ -277,7 +288,7 @@ export const visualizationSchema = z.object({
 		.object({
 			showKey: z.boolean(),
 			style: z.string().optional(),
-			strategy: z.string().optional(),
+			strategy: z.nativeEnum(LegendStrategy).optional(),
 			set: z
 				.object({
 					id: z.string(),
@@ -333,4 +344,16 @@ const supportedVisualizations = ["MAP", "chart", "table"] as const;
 
 export type SupportedVisualization = (typeof supportedVisualizations)[number];
 
+export const yearOverYearVisualizationSchema = visualizationSchema.extend({
+	type: z.enum([
+		VisualizationChartType.YEAR_OVER_YEAR_COLUMN,
+		VisualizationChartType.YEAR_OVER_YEAR_LINE,
+	]),
+	yearlySeries: z.array(z.string()),
+	relativePeriods: z.record(z.string(), z.boolean()),
+});
+
 export type VisualizationConfig = z.infer<typeof visualizationSchema>;
+export type YearOverYearVisualizationConfig = z.infer<
+	typeof yearOverYearVisualizationSchema
+>;